obd/method_stubs.py
import tkinter as tk
from tkinter import ttk, messagebox, filedialog
import customtkinter as ctk
import obd
import serial
import time
import functools
import serial.tools.list_ports
import threading
import time
import json
import datetime
import csv
import matplotlib.pyplot as plt
from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
import numpy as np
from typing import Dict, List, Optional
import logging
import os
import asyncio
import bleak
from bleak import BleakClient, BleakScanner

# Method stubs for EnhancedOBDMonitor class


class EnhancedOBDMonitorStubs:
    def __init__(self):
        pass

    # Add stubs for all missing methods
    def on_closing(self):
        """Handle window close event: save config and cleanup."""
        pass

    def update_session_timer(self):
        """Update session timer display"""
        pass

    def start_monitoring(self):
        """Start OBD monitoring"""
        pass

    def stop_monitoring(self):
        """Stop OBD monitoring"""
        pass

    def export_data(self):
        """Export monitoring data to CSV file"""
        pass

    def update_connection_status(self, connected):
        """Update connection status in UI"""
        pass

    def get_available_pids(self):
        """Get list of available PIDs from the vehicle"""
        pass

    def setup_enhanced_logs_tab(self):
        """Setup enhanced logs tab"""
        pass

    def setup_settings_tab(self):
        """Setup settings tab"""
        pass

    def show_add_pid_dialog(self):
        """Show dialog to add PID to dashboard"""
        pass

    def create_gauge_row(self):
        """Create a new row frame for gauges"""
        pass

    def add_gauge_to_dashboard(self, title, key, min_val, max_val, unit, color):
        """Add a gauge to the dashboard"""
        pass

    def filter_pids(self, event=None):
        """Filter PIDs based on search term"""
        pass

    def export_pids_data(self):
        """Export current PID data to CSV file"""
        pass

    def start_charting(self):
        """Start real-time charting for selected PID"""
        pass

    def stop_charting(self):
        """Stop real-time charting"""
        pass

    def refresh_dtcs(self):
        """Refresh DTC data from vehicle"""
        pass

    def export_dtcs(self):
        """Export DTC data to text file"""
        pass
