obd/requirements.txt
customtkinter>=5.2.0
obd>=0.7.1
pyserial>=3.5
matplotlib>=3.5.0
numpy>=1.21.0
bleak>=0.19.0
pandas>=1.3.0
